Goals: Reduce the rendering/processing time so that scrolling is not too choppy
Solved.
Added call to unload dropped files to clear the dropped files cache, in turn reducing the overhead.
Also added a check that we're not processing the same file twice. It should not be necessary anymore as files are only processed once but should reduce the cost for large files